add ( detail ) : add ProductDetailV4
