+++ lib/constants/colors/product_v4_colors.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 class ProductV4Colors {
   static const int primaryColor = 0xff0D3A5C;
   static const int primaryColorDark = 0xff0D3A5C;
@@ -70,72 +72,83 @@ class ProductContainerStyles {
 }
 
 final product = ProductContainerStyles(
-    backgroundContainerDark: 0xff1B1B1B,
-    backgroundContainerLight: 0xffE9FAFF,
-    imageProduct: "🏢",
-    titleText: "Producto de inversión a Plazo Fijo",
-    minimumText: "1.000",
-    profitabilityText: "19",
-    titleDark: 0xffFFFFFF,
-    titleLight: 0xff0D3A5C,
-    minimumDark: 0xff0D3A5C,
-    minimumLight: 0xffBBF0FF,
-    profitabilityDark: 0xffB5FF8A,
-    profitabilityLight: 0xffD2FDBA,
-    isSoles: true,
-    uuid: "1",
-    buttonBackDark: 0xffA2E6FA,
-    buttonBackLight: 0xff0D3A5C,
-    buttonTextDark: 0xff0D3A5C,
-    buttonTextLight: 0xffFFFFFF,
-    textDark: 0xff000000,
-    textLight: 0xff000000,
-    minimunTextColorDark: 0xff000000,
-    minimumTextColorLight: 0xff000000);
+  backgroundContainerDark: 0xff1B1B1B,
+  backgroundContainerLight: 0xffE9FAFF,
+  imageProduct: "🏢",
+  titleText: "Producto de inversión a Plazo Fijo",
+  minimumText: "1.000",
+  profitabilityText: "19",
+  titleDark: 0xffFFFFFF,
+  titleLight: 0xff0D3A5C,
+  minimumDark: 0xff0D3A5C,
+  minimumLight: 0xffBBF0FF,
+  profitabilityDark: 0xffB5FF8A,
+  profitabilityLight: 0xffD2FDBA,
+  isSoles: true,
+  uuid: "1",
+  buttonBackDark: 0xffA2E6FA,
+  buttonBackLight: 0xff0D3A5C,
+  buttonTextDark: 0xff0D3A5C,
+  buttonTextLight: 0xffFFFFFF,
+  textDark: 0xff000000,
+  textLight: 0xff000000,
+  minimunTextColorDark: 0xffFFFFFF,
+  minimumTextColorLight: 0xff000000,
+);
 
 final product2 = ProductContainerStyles(
-    backgroundContainerDark: 0xff1B1B1B,
-    backgroundContainerLight: 0xffE9FAFF,
-    imageProduct: "🏡",
-    titleText: "Producto de inversión con Garantía Inmobiliaria",
-    minimumText: "50.000",
-    profitabilityText: "16",
-    titleDark: 0xffFFFFFF,
-    titleLight: 0xff0D3A5C,
-    minimumDark: 0xff0D3A5C,
-    minimumLight: 0xffBBF0FF,
-    profitabilityDark: 0xffB5FF8A,
-    profitabilityLight: 0xffD2FDBA,
-    isSoles: true,
-    uuid: "2",
-    buttonBackDark: 0xffA2E6FA,
-    buttonBackLight: 0xff0D3A5C,
-    buttonTextDark: 0xff0D3A5C,
-    buttonTextLight: 0xffFFFFFF,
-    textDark: 0xff000000,
-    textLight: 0xff000000,
-    minimunTextColorDark: 0xff000000,
-    minimumTextColorLight: 0xff000000);
+  backgroundContainerDark: 0xff1B1B1B,
+  backgroundContainerLight: 0xffE9FAFF,
+  imageProduct: "🏡",
+  titleText: "Producto de inversión con Garantía Inmobiliaria",
+  minimumText: "50.000",
+  profitabilityText: "16",
+  titleDark: 0xffFFFFFF,
+  titleLight: 0xff0D3A5C,
+  minimumDark: 0xff0D3A5C,
+  minimumLight: 0xffBBF0FF,
+  profitabilityDark: 0xffB5FF8A,
+  profitabilityLight: 0xffD2FDBA,
+  isSoles: true,
+  uuid: "2",
+  buttonBackDark: 0xffA2E6FA,
+  buttonBackLight: 0xff0D3A5C,
+  buttonTextDark: 0xff0D3A5C,
+  buttonTextLight: 0xffFFFFFF,
+  textDark: 0xff000000,
+  textLight: 0xff000000,
+  minimunTextColorDark: 0xffFFFFFF,
+  minimumTextColorLight: 0xff000000,
+);
 final product3 = ProductContainerStyles(
-    backgroundContainerDark: 0xff1B1B1B,
-    backgroundContainerLight: 0xffE9FAFF,
-    imageProduct: "🏢",
-    titleText: "Producto de inversión a Plazo Fijo",
-    minimumText: "1.000",
-    profitabilityText: "17",
-    titleDark: 0xffFFFFFF,
-    titleLight: 0xff0D3A5C,
-    minimumDark: 0xff0D3A5C,
-    minimumLight: 0xffBBF0FF,
-    profitabilityDark: 0xffB5FF8A,
-    profitabilityLight: 0xffD2FDBA,
-    isSoles: false,
-    uuid: "3",
-    buttonBackDark: 0xffA2E6FA,
-    buttonBackLight: 0xff0D3A5C,
-    buttonTextDark: 0xff0D3A5C,
-    buttonTextLight: 0xffFFFFFF,
-    textDark: 0xff000000,
-    textLight: 0xff000000,
-    minimunTextColorDark: 0xff000000,
-    minimumTextColorLight: 0xff000000);
+  backgroundContainerDark: 0xff1B1B1B,
+  backgroundContainerLight: 0xffE9FAFF,
+  imageProduct: "🏢",
+  titleText: "Producto de inversión a Plazo Fijo",
+  minimumText: "1.000",
+  profitabilityText: "17",
+  titleDark: 0xffFFFFFF,
+  titleLight: 0xff0D3A5C,
+  minimumDark: 0xff0D3A5C,
+  minimumLight: 0xffBBF0FF,
+  profitabilityDark: 0xffB5FF8A,
+  profitabilityLight: 0xffD2FDBA,
+  isSoles: false,
+  uuid: "3",
+  buttonBackDark: 0xffA2E6FA,
+  buttonBackLight: 0xff0D3A5C,
+  buttonTextDark: 0xff0D3A5C,
+  buttonTextLight: 0xffFFFFFF,
+  textDark: 0xff000000,
+  textLight: 0xff000000,
+  minimunTextColorDark: 0xffFFFFFF,
+  minimumTextColorLight: 0xff000000,
+);
+
+class ChartData {
+  final String category;
+  final double value;
+  final Color color;
+
+  ChartData(this.category, this.value, this.color);
+}
